Finalização das demandas para ficar usável
esta versão já está sendo usada em PRD

+++ src/spark/template/freemarker/grupolistar.ftl
@@ -2,6 +2,6 @@
 <h1>Seus grupos de gasto</h1>
 <#list (gruposGasto) as g>
 ${g.nome} ${g.id} 
-<a href='/grupogasto/editar/${g.id}'>A</a> <a href='/grupogasto/excluir/${g.id}'>X</a><br/>
+<a href='/grupogasto/editar/${g.id}'>A</a> <a href='/grupogasto/excluir/${g.id}' class="confirmation">X</a><br/>
 </#list>
 <#include "footer.ftl">
