+++ src/spark/template/freemarker/periodolistar.ftl
@@ -2,6 +2,6 @@
 <h1>Seus períodos de contabilização</h1>
 <#list (lista) as g>
 ${g.idPeriodo}: de ${g.dataInicio?date?string("dd/MM/yyyy")} a ${g.dataFim?date?string("dd/MM/yyyy")} 
-<a href='/periodo/editar/${g.idPeriodo}'>A</a> <a href='/periodo/excluir/${g.idPeriodo}'>X</a><br/>
+<a href='/periodo/editar/${g.idPeriodo}'>A</a> <a href='/periodo/excluir/${g.idPeriodo}' class="confirmation">X</a><br/>
 </#list>
 <#include "footer.ftl">
